Rename: 패키지 이름 ui -> common 변경
