+++ feature/main/src/main/java/kr/co/main/my/community/bookmark/MyPageBookmarkRoute.kt
@@ -29,7 +29,7 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kr.co.main.R
-import kr.co.main.ui.DreamMainPostCard
+import kr.co.main.common.DreamMainPostCard
 import kr.co.ui.ext.scaffoldBackground
 import kr.co.ui.theme.NBDreamTheme
 import kr.co.ui.theme.colors
